Refactor to move alloc stuff to corealloc.c (3/8)
In this commit we move over the __prov_track_alloc() and
__prov_track_dealloc() functions from schedule.c into corealloc.c. We
also rename them to more meaningful names in the process.
